finished organazing all the commands ported from java so far
